requirements.txt
firecrawl-py
python-dotenv
pandas
beautifulsoup4
